ASoC: tegra-alt: Manually kick off RT5658 jack detection feature
As the JD3 pin is connected to the GND only, the internal
Jack Detection function cannot generate a normal IRQ from
GPIO1 in order to run the jack detection routine via the
interrupt callback function.

So this could be treated as a work around to trigger the
jack detection manually.

Bug 1736796

Change-Id: Iee09cb41b976fdeae458fb081c84407fd84d02a2
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
Reviewed-on: http://git-master/r/1207328
GVS: Gerrit_Virtual_Submit
Reviewed-by: Ravindra Lokhande <rlokhande@nvidia.com>