arm64: dts: t194: remove mem param from commandline
It is unnecessary to specify the mem param in kernel commandline in
addition to /memory node.

In case of frontdoor, BL considers the commandline in DTS as override
for BL settings, hence keeping mem param limits the usable memory whereas
the actual size is advertised by BL in /memory node.

In case of backdooor, the sysimg packer script takes care of
overwriting the kernel commandline as required.

Bug 1805920

Change-Id: Ic627fe63c64a03045babe3780fe47b16de546766
Signed-off-by: Sukanto Ghosh <sukantog@nvidia.com>
Reviewed-on: http://git-master/r/1212673
GVS: Gerrit_Virtual_Submit
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>