denver: hardwood: use device tree to find IRQ number
This driver is currently disabled in kernel-4.4.
When attempting to reenable it, there was an IRQ
conflict, because the hardwood driver had a hardcoded
IRQ that conflicted with a USB device in the device
tree. This restructures the driver to get its IRQ
from the device tree, so we can avoid this problem
in the future.

bug 1799795
Bug 200325746

Change-Id: Ia6f3b340a423edd97135b69d5f1a6154851d974a
Signed-off-by: Rich Wiley <rwiley@nvidia.com>
Reviewed-on: http://git-master/r/1218926
GVS: Gerrit_Virtual_Submit
Reviewed-by: Alexander Van Brunt <avanbrunt@nvidia.com>
(cherry picked from commit 1bc0602fff6b4719d44e72bd7fb736c4fcbf9017)
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1516722
Reviewed-by: Sachin Nikam <snikam@nvidia.com>