video: tegra: host: pva: Mailbox updates
This patch reworkes the mailbox interface to support request-response
model.

First, the patch removes an unused status poll function.

Second, interrupt handling is rewritten: Interrupts are handled
based on ISR_PENDING flags in the status registers. If AISR
exception or SWUART interrupt is received, the exception is
written into the kernel log.

Third, PVA interrupts are also reworked: The status register is
always cleared, however, status information is stored only if there
is a pending request. Once the status data is handled, the waiter
is signalled.

Fourth, A new function, pva_mailbox_send_cmd_sync() is added to
submit a request and wait for a response.

JIRA PVA-175

Change-Id: I6cc80cec3350f9cdd0a9b772cb0b8c749b482217
Signed-off-by: Vinod G <vinodg@nvidia.com>
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/1222215
GVS: Gerrit_Virtual_Submit