ARM64: dts: tegra210: add clock for Tegra timer
Add clock and refine interrupt property for the Tegra timer.
Note that two drivers are using the node. One is Tegra timer to replace
the arch timer and uses as system timer, the compatible string is
"nvidia,tegra210-timer". The other one is Tegra watchdog timer, the
compatible string is "nvidia,tegra30-timer".

Bug 200237195

Change-Id: Id326e38c15e449d39153e85482577c329d34612f
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Reviewed-on: http://git-master/r/1231797
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>