video: tegra: host: Use mark_unused for syncpt protection
To make syncpts unavailable to accesses from wrong channels/
when unused nvhost sets their channel to a nonexisting one
when they are not in use. This is currently done in the .reset
callback, which works except when the syncpt is reset due
to a submit failing. In that case the reset callback is invoked
to reset the syncpoint's value to the expected one after
the submission, but due to the protection code the syncpt
is rendered unusable for subsequent submissions on the channel.

Move the protection code to .mark_unused. This in conjunction
with modifications to host1x initialization code fixes this
behavior.

Jira HOSTX-116

Change-Id: I95e1a9d6699038a97c3ae8233d391376cc42ff90
Signed-off-by: Mikko Perttunen <mperttunen@nvidia.com>
Reviewed-on: http://git-master/r/1228979
Reviewed-by: Arto Merilainen <amerilainen@nvidia.com>