sched_clock: Change sched_clock epoch to POR time
Sched_clock is taking the time of clock registration
as the epoch(start of time), So that the time starts from
zero for the sched_clock.
Convert it to consider the time of POR(power on reset) as the
epoch time. This is done for printk to have the timings from board
bootup.

Bug 1810577
Bug 200396169

Change-Id: If3e59c1caf5a27b1570e9be3c811359a1310d4db
Signed-off-by: Virupax.Sadashivpetimath <virupaxs@nvidia.com>
Reviewed-on: http://git-master/r/1230621
(cherry picked from commit 782f49e27cc17049e153eb33bca27990589d0270)
Signed-off-by: dmitry pervushin <dpervushin@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1679577
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: Vipin Kumar <vipink@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-by: Bo Yan <byan@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>