video: tegra: fb: Check for errors when setting up videomode
The return code from the function tegra_dc_to_fb_videomode() is not
being checked. This function can return an error if the videomode has
not been setup correctly. This was happening on Tegra210 Jetson TX1
when enabling CONFIG_FRAMEBUFFER_CONSOLE because no odm-data was
present in the device-tree image (causing fb_console_mapped() to
return false).

If tegra_dc_to_fb_videomode() returns an error then this may lead to
a kernel crash, because x and y resolutions will not be configured
correctly and this can lead memory corruption when the frame-buffer
is initialised. Therefore, check the return value from
tegra_dc_to_fb_videomode() to avoid an potential crashes.

Bug 200237207

Change-Id: Ic332171d81af7e3a6fb80bb07c129243f92522d8
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: http://git-master/r/1233343
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>