of: plugin-manager: Add support of alpha character FabID
Add support of the FabID containing the alphabetical character
also. This will help to extend the FabID to large number of
combination.

Treat caps and small letter as same. Make valid FabID for
following values: '0' to '9', 'A' to 'Z' and 'a' to 'z'.

bug 1827258

Change-Id: I004706a74e0ddd58fe7a535e784dfa62e5c62946
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/1242108
(cherry picked from commit 1652ccdbc6d29a7a8335e003dd9ca54de9929775)
Signed-off-by: Venkat Reddy Talla <vreddytalla@nvidia.com>