tegra-se: Fix input/output buf linked list issue
Hardware expects each element size in Input/output
buffer lined-list is of eight bytes. Currently in
driver it is not eight bytes. Hence when input/output data
is in multiple buffers there is mismatch in the result.

Fixed it by making each element size in linked-list to
be of eight bytes.

Also set coherent_dma_mask to 32 bit as SE HW supports
32 bit address.

Bug 200237277

Change-Id: I3f561903b20b8ecf4dceac0e61ab1c9f413af007
Signed-off-by: Mallikarjun Kasoju <mkasoju@nvidia.com>
Reviewed-on: http://git-master/r/1246386
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Signed-off-by: Shravani Dingari <shravanid@nvidia.com>