video: tegra: host: t194: Fix PVA clock names
PVA clock names currently include the PVA ID and their names
are not reflecting the names in the device tree (e.g. pva0 vs axi).
Having the ID makes reusage of the code harder and discrepancy
in the naming makes it unclear which of the PVA clocks is being
enabled.

This change updates the clock names to not have PVA ID and keep
the naming convention same as in the BPMP ABI.

JIRA HOSTX-141

Change-Id: Iebf4b65b87e8b27600839ec1ed02b67543edf872
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/1248908
Reviewed-by: Shridhar Rasal <srasal@nvidia.com>
GVS: Gerrit_Virtual_Submit