video: tegra: Re-introduce CONFIG_FRAMEBUFFER_CONSOLE dependency
Commit f0b131cc3a6d ("video: tegra: remove CONFIG_FRAMEBUFFER_CONSOLE")
removed the compile time if-statements for CONFIG_FRAMEBUFFER_CONSOLE
to pave the way for having a single kernel configuration for the various
linux distributions (eg, android, l4t, etc). However, by removing this
it means there needs to be a boot-time method for detecting whether the
frame-buffers should be registered with fbcon. Although there is such a
method currently (through the odm-data that is dynamically inserted into
the device-tree blob), this is not maintainable because this is not
something that can be upstreamed. In the short-term it is simpler to
revert to making the code compile time dependent on
CONFIG_FRAMEBUFFER_CONSOLE.

Change-Id: I903fe8cb7ce8c0f860dcd0e97528c0ed0a042085
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: http://git-master/r/1250553
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>