t18x: add mce debugfs node for setting edbgreq on MCA
This patch adds access to a debugfs node that,
via kernel SMC call to ATF and and ATF ARI call to MCE,
enables the EDBGREQ feature in the CPU.

When enabled this feature will put the CPU in the debug
state instead of vectoring to sw (ie the MCA handler)
upon receiving an asynchronous abort signal. This allows
debug very close to the point of the actual error
occurring.

bug 1839113

Change-Id: I2c211e30493d43b0d0a03fa041744b88b090af4f
Signed-off-by: Rich Wiley <rwiley@nvidia.com>
Reviewed-on: http://git-master/r/1271253
Reviewed-by: Alexander Van Brunt <avanbrunt@nvidia.com>
GVS: Gerrit_Virtual_Submit