camera: rtcpu: Use correct rtpm pair func
pm_runtime_get_sync increments the ref-count
before resuming the device, so to set the device
back to suspend/idle state, need to use correct
rtpm pair function which also uses the ref-count,
the pm_runtime_put_sync, it decrements the ref-count
and if its 0, sets the device to idle.

Using pm_runtime_idle directly with get_sync, does
not work, as it doesn't deal with ref-count, so
if ref-count is non-zero, it simply returns as no-op.

Bug 1779439
Jira CRTC-692

Change-Id: I2604aa1de198eecbb42a6161a63d1637a323ba43
Signed-off-by: Sudhir Vyas <svyas@nvidia.com>
Reviewed-on: http://git-master/r/1265094
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Pekka Pessi <ppessi@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Mika Liljeberg <mliljeberg@nvidia.com>
Reviewed-by: Jihoon Bang <jbang@nvidia.com>