t210: jetson: dts: default rates for i2s and dmic
The patch is to avoid initial clk_enable() failure on i2s clock.
Failure is seen because of initial clock/divider settings and the
sequence in which audio clocks are enabled. This fix ensures i2s
clock to be initialized with known good value. Similar change is
done for dmic as well.

Bug 1810165
Bug 200253201

Change-Id: Ie6293e766199acfc128dbb7bd9f766a2664216d6
Reviewed-on: http://git-master/r/1265673
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>