padctl: t186: set VBUS dir in VBUS pad protection
In T186 we have VREG_DIR[12:11] bits for setting OTG port VBUS direction
for sink or source. When the OTG port is A-device the VBUS is always
source (supplying VBUS), and when the OTG port is B-device the VBUS is
always sink. However, when A-device acts as a-peripheral (still
supplying VBUS but acts as an USB device) following OTG2.0 HNP, VBUS
OVERRIDE bit is also set. It is hence unreliable to use VBUS OVERRIDE bit
to determine when OTG port is currently acting as VBUS source or sink.

In other words, padctl driver has no knowledge when the OTG port is source
or sink. Hence we add a new parameter to the pad protection API:
tegra_phy_xusb_utmi_pad_set_protection_level() for drivers to set the VBUS
direction while setting pad protection.

Bug 200261283

Change-Id: I9472f60041b4743904011c0baf812a12c3216053
Signed-off-by: Mark Kuo <mkuo@nvidia.com>
Reviewed-on: http://git-master/r/1269352
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Ashutosh Jha <ajha@nvidia.com>