crypto: get rid of chip specific defconfig
remove chip-specific guards that were added
to build some code only for a particulr soc.
The defconfig change that is needed for this change
is @ http://git-master/r/#/c/1271451/
Also T30 RNG specific code is removed in this change
as this code is no more used for current or future
chips.

Bug 200257391

Change-Id: I81e712abdc97a932a645c8fe246654a0c6c67213
Signed-off-by: Shravani Dingari <shravanid@nvidia.com>
Reviewed-on: http://git-master/r/1271448
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Mallikarjun Kasoju <mkasoju@nvidia.com>
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>