video: tegra: host: Determine chip at runtime
Due to kernel image unification, compile time defines
should not be used to determine the runtime platform.
Therefore, determine it at runtime.

The proper way to do this would be to add SoC info
structures to device tree match data, but due to the
way the nvhost driver is structured, this is almost
impossible. Therefore this patch implements some helpers
to get the current chip version using the chip ID
functions and revises code to use those.

Bug 200261532

Change-Id: Idc67828eeb183bdc79700924a0c378800a56054d
Signed-off-by: Mikko Perttunen <mperttunen@nvidia.com>
Reviewed-on: http://git-master/r/1275671
GVS: Gerrit_Virtual_Submit
Reviewed-by: Arto Merilainen <amerilainen@nvidia.com>