trusty-ipc: Rework tipc_chan_destroy
It seems unnecessary to dynamically detach vds pointer in
tipc_chan_destroy. Instead, it is cleaner to keep it around
until channel object itself is actually freed. This eliminates
the need to dynamically check presence of that pointer as it
is should always be present if tipc_chan exists.

Change-Id: Id9f7181aa4934c1e5d2eaf6d245fa97654451212
Signed-off-by: Michael Ryleev <gmar@google.com>
Reviewed-on: https://git-master.nvidia.com/r/1679988
Tested-by: Stephen Wolfe <swolfe@nvidia.com>
Reviewed-by: Mihir Joshi <mihirj@nvidia.com>
Reviewed-by: Mahesh Lagadapati <mlagadapati@nvidia.com>
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>