video: tegra: unpowergate dc during resume
i2c errors are seen when hpd worker tries to read EDID
after resuming from LP0. This issue is seen if no userspace
is running and only HDMI is active.

When resuming, responsibility of enabling DC is on the
userspace entity so that it can get DC enabled according to
its preferences.
But when no userspace entity is active, dc stays powergated
which leads to failure in EDID read by hpd worker.

Hence unpowergate DC in tegra_dc_resume() if no userspace entity
is active.

bug 200258075

Change-Id: Icda93a60b5b973c7a69c2decece204f41a14c9ab
Signed-off-by: Naveen Kumar S <nkumars@nvidia.com>
Reviewed-on: http://git-master/r/1277154
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>