video: tegra: host: Fix .can_powergate flag
Earlier .can_powergate flag behavior relied on generic power domains
behavior: If domain was powered, also the client state was assumed
to be correct. When a domain was powered down, also the client state
was stored.

However, in K4.4 this relationship isn't correct but the pm runtime
callbacks for saving and restoring device state may be called
regardless of the power domain state. Since .can_powergate is
controlling the domain state, the client currently may still be
actively turned on and off even if .can_powergate flag is marked
as false.

This change modifies code to take explicitly a pm_runtime reference
on the device if .can_powergate flag is set to false. This should
ensure that the device power domain is kept powered and that the
device state is not saved/restored actively during runtime.

If there is a suspend-resume -cycle, the device state is stored
before suspending and restored after resuming.

JIRA PVA-165

Change-Id: Ib17ba50f61bf4381d5009f9eed705c8cc289cc03
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/1278748
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Mikko Perttunen <mperttunen@nvidia.com>
GVS: Gerrit_Virtual_Submit