nvdisp: pool ordering should take account of OLUT
For a given head, the output LUT shares the same precomp
pipe and FB request port as the corresponding cursor pipe.
As a result, for a usecase where cursor is disabled but
output LUT is enabled, the cursor ihub settings - such
as assigned mempool entries - would still be programmed
for the given head.

This patch updates the function that's responsible
for generating the mempool programming order so that
it no longer assumes that no cursor mempool entries are
assigned if only cursor is disabled.

Bug 1856079
Bug 1847292
JIRA TDS-1570

Change-Id: Ib29b49f176ccb009a50cdf0d52442f20eab0256d
Signed-off-by: Shu Zhong <shuz@nvidia.com>
Reviewed-on: http://git-master/r/1278833
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>