tegra-se: fixed probe call failure.
Device ids are given in device driver structure
so that they will be matched against the entries
in DT.If a match is found, driver will be called.

The SE driver probe for T214 FPGA is not getting
called because Device id entries in device driver
declaration Is not initialized as it is guarded
with CONFIG_PM define.

Fixed the issue by taking out the open firmware
table (of_match_table) entry in device driver
declaration  from CONFIG_PM

Change-Id: I26d4c05fb1bdac26c63732ca83e765d761b97f87
Signed-off-by: kondurip <kondurip@nvidia.com>
Reviewed-on: http://git-master/r/1280252
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>