tegra: mc: t18x: disp is passing emc freq
Initially it was assumed that there was a bug in the display driver. It
was thought that the display driver may be passing DRAM frequency to the
LA/PTSA driver through API calls. This is not the case. Display is
correctly passing the EMC frequency to the LA/PTSA driver. Remove
comment about this issue from the LA/PTSA driver.

Bug 1764786

Change-Id: If868b6e4d6f039272cb8ac9f9fc51d1c91fa5f76
Signed-off-by: Adeel Raza <araza@nvidia.com>
Reviewed-on: http://git-master/r/1281026
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Shu Zhong <shuz@nvidia.com>
GVS: Gerrit_Virtual_Submit