tegra: mc: t18x: disp is passing emc freq
Initially it was assumed that there was a bug in the display driver. It
was thought that the display driver may be passing DRAM frequency to the
LA/PTSA driver through API calls. This is not the case. Display is
correctly passing the EMC frequency to the LA/PTSA driver. Remove
comment about this issue from the LA/PTSA driver.

Bug 1764786

Change-Id: I6bcf384f617c1ff23afa226562adb3d11ca8a6f6
Signed-off-by: Adeel Raza <araza@nvidia.com>
Reviewed-on: http://git-master/r/1281026
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Shu Zhong <shuz@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-on: http://git-master/r/1469896
Reviewed-by: Krishna Reddy <vdumpa@nvidia.com>
Tested-by: Krishna Reddy <vdumpa@nvidia.com>