video: tegra: dc: export DC reg access routines
MODS display adds private header of display drivers and uses inline
functions for register access. These functions further use routines
from CCF which are not exported which leads to compilation errors.
Lack of formal SW interface between MODS and display driver is the
root cause for this kind of issues. While that interface is being
worked on, in the interim add wrapper exported function in display
driver for MODS to use.

bug 1854882

Change-Id: I6ae528880be2750fcc652c4e8c22efb91c71ee0f
Signed-off-by: Ujwal Patel <ujwalp@nvidia.com>
Reviewed-on: http://git-master/r/1281399
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>