nvdisp: filter IMP results on head active field
Even if the driver detects that a head is currently enabled,
ignore its corresponding IMP results if the head active field
isn't set. Userspace is responsible for setting this field
during PROPOSE for each active head that was fed into the IMP
algorithm. The above check is required since the PROPOSE and
FLIP phases aren't necessarily atomic with respect to each other.

Bug 1847292
JIRA TDS-1570

Change-Id: I30313cd455007424215486905a353af72635aac4
Signed-off-by: Shu Zhong <shuz@nvidia.com>
Reviewed-on: http://git-master/r/1282580
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>