firmware: tegra: fix return value from store fns
bpmp_trace_enable_store() and bpmp_trace_disable_store() are now
returning the updated trace mask value instead of  0 or an error code.
This results in a write() system call to return wrong values to the
caller. Fix this by returning an error code or just 0 if there were no
errors.

Bug 200252593

Change-Id: Ifad23427c679a6fa9f0a34cb70a596469b0a8b94
Signed-off-by: Sivaram Nair <sivaramn@nvidia.com>
Reviewed-on: http://git-master/r/1283087