video: tegra: host: Don't lock in power_{on,off} handlers
In kernel 4.9, power domains are turned on before probe is
run for a device that will be attached to them. Therefore
we cannot use the lock in device data that is only
initialized in the device's probe. The lock is unnecessary
anyway, since genpd already locks these callbacks and
checking can_powergate is unnecessary, as a device with
can_powergate set to false will always have a runtime PM
reference to it.

JIRA HOSTX-156
Bug 1852328

Change-Id: I613b7794c742323cf69350f3ed919e6cb5a12b1f
Signed-off-by: Mikko Perttunen <mperttunen@nvidia.com>
Reviewed-on: http://git-master/r/1284185
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>